feat: refactor duration utils
