+++ src/lib/utils.ts
@@ -22,29 +22,6 @@ export function sanitizeViewTransitionName(name: string | undefined): string | u
 	return sanitized || 'tag';
 }
 
-export const DURATION_BUCKETS = [15, 30, 45, 60, 90, 120, 180] as const;
-export type DurationBucket = (typeof DURATION_BUCKETS)[number];
-
-export function formatDuration(minutes: number | null | undefined): string | null {
-	if (minutes == null) return null;
-	switch (minutes) {
-		case 15:
-		case 30:
-		case 45:
-			return `${minutes} min`;
-		case 60:
-			return '1 h';
-		case 90:
-			return '1.5 h';
-		case 120:
-			return '2 h';
-		case 180:
-			return '3+ h';
-		default:
-			return `${minutes} min`;
-	}
-}
-
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type WithoutChild<T> = T extends { child?: any } ? Omit<T, 'child'> : T;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
